+++ CHANGELOG.md
@@ -11,6 +11,9 @@ accidentally triggering the load of a previous DB version.**
 * #4169 Add support for chunk exclusion on DELETE to PG14
 * #4209 Add support for chunk exclusion on UPDATE to PG14
 
+**Bugfixes**
+* #4225 Fix TRUNCATE error as non-owner on hypertable
+
 ## 2.6.1 (2022-04-11)
 This release is patch release. We recommend that you upgrade at the next available opportunity.